requirements.txt
PyQt6==6.4.2
anthropic==0.15.0
pynput==1.7.6
python-dotenv==1.0.1 